+++ src/function/matrix/zeros.js
@@ -2,7 +2,9 @@ module.exports = function (math) {
   var util = require('../../util/index.js'),
 
       Matrix = require('../../type/Matrix.js'),
-      collection = require('../../type/collection.js');
+      collection = require('../../type/collection.js'),
+
+      array = util.array;
 
   /**
    * create a matrix filled with zeros
@@ -13,21 +15,22 @@ module.exports = function (math) {
    *     zeros([m, n, p, ...])
    *
    * @param {...Number | Array} size
-   * @return {Matrix} matrix
+   * @return {Array | Matrix | Number} matrix
    */
   math.zeros = function zeros (size) {
     var args = collection.argsToArray(arguments);
+    var asMatrix = (size instanceof Matrix);
 
     if (args.length == 0) {
-      args = [1, 1];
+      // output a scalar
+      return 0;
     }
-    else if (args.length == 1) {
-      args[1] = args[0];
+    else {
+      // output an array or matrix
+      var res = [];
+      var defaultValue = 0;
+      array.resize(res, args, defaultValue);
+      return asMatrix ? new Matrix(res) : res;
     }
-
-    // create and size the matrix
-    var matrix = new Matrix();
-    matrix.resize(args);
-    return matrix;
   };
 };